# src/main/java/com/ejemplo/demo/api/controller/SaludoController.java
package com.ejemplo.demo.api.controller;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import com.ejemplo.demo.domain.service.SaludoService;
import org.springframework.web.bind.annotation.RequestParam;
import jakarta.validation.Valid;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import com.ejemplo.demo.generated.api.WorkshopApi;
//import com.ejemplo.demo.generated.model.SaludoRequest;
//import com.ejemplo.demo.generated.model.SaludoResponse;
import com.ejemplo.demo.api.dto.SaludoRequest;
import com.ejemplo.demo.api.dto.SaludoResponse;

import java.util.Map;
import java.util.List;

@RestController
@RequestMapping("/api/v1")
public class SaludoController implements WorkshopApi {
	
	@Override
	public ResponseEntity<List<String>> listarSaludos() {

	    return ResponseEntity.ok(
	            List.of("Hola", "Bienvenido")
	    );
	}
	
	
	@Override
	public ResponseEntity<SaludoResponse> crearSaludo(
	        @Valid @RequestBody SaludoRequest request
	) {

	    return ResponseEntity.ok(
	            saludoService.crearSaludo(request.nombre())
	    );
	}

    @GetMapping
    public ResponseEntity<Map<String, String>> health() {
        return ResponseEntity.ok(Map.of(
                "estado", "ok",
                "mensaje", "Workshop Spring Boot activo"
        ));
    }

    
   // ===========================================
    //PASO 2: DESCOMENTA este bloque y prueba GET
   // ===========================================

  //  1) Descomenta imports:
     

    //2) Descomenta el campo y constructor:

    private final SaludoService saludoService;

    public SaludoController(SaludoService saludoService) {
        this.saludoService = saludoService;
    }

    //3) Descomenta este endpoint:
    
    /*
    @GetMapping("/saludos")
    public ResponseEntity<SaludoResponse> saludar(
            @RequestParam(defaultValue = "Mundo") String nombre
    ) {
        return ResponseEntity.ok(saludoService.crearSaludo(nombre));
    }
    

    /*
    ============================================
    PASO 3: DESCOMENTA este bloque y prueba POST
    ============================================

    1) Descomenta imports:
       - 

    2) Descomenta este endpoint:
     
    @PostMapping("/saludos")
    public ResponseEntity<SaludoResponse> saludarPost(@Valid @RequestBody SaludoRequest request) {
        return ResponseEntity.ok(saludoService.crearSaludo(request.nombre()));
    }
    */
}