# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Spring Boot API Demo
  version: 1.0.0
  description: API REST para Programacion III
servers:
  - url: http://localhost:8080
...
paths:
...
  /saludos:
    get:
      tags:
        - workshop

      summary: Listar saludos
      operationId: listarSaludos

      responses:
        '200':
          description: Lista de saludos
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

    post:
      tags:
        - workshop

      summary: Crear saludo
      operationId: crearSaludo

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaludoRequest'

      responses:
        '200':
          description: Saludo creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaludoResponse'
  /api/v1/simulaciones/prestamo:
    post:
      tags:
        - simulaciones

      summary: Simular prestamo
      operationId: simularPrestamo

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrestamoRequest'

      responses:
        '200':
          description: Resultado simulacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrestamoResponse'
...
components:
  schemas:
    SaludoRequest:
      type: object

      properties:
        mensaje:
          type: string
    SaludoResponse:
      type: object

      properties:
        respuesta:
          type: string
    PrestamoRequest:
      type: object

      properties:
        monto:
          type: number
          format: double

        tasaAnual:
          type: number
          format: double

        meses:
          type: integer
    PrestamoResponse:
      type: object

      properties:
        cuotaMensual:
          type: number
          format: double

        interesTotal:
          type: number
          format: double

        totalPagar:
          type: number
          format: double